Added view functionaliaty for sequences and generic views.
This commit includes a lot of changes to support generic views for
processes. It also includes "selections" to determine which stock to
show based on GET query paramaters. Finally it uses selections to
imlement a sequencer which makes it easy to flip through the objects in
a stock one by one.